chore: rename app.tsx
