+++ src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App.tsx'
+import App from './app'
 import './main.css'
 
 createRoot(document.getElementById('root')!).render(
